[ClearlyDefined] New badge
